backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
python-dotenv==1.0.0
PyJWT==2.8.0
Werkzeug==3.0.1
bcrypt==4.1.2
pytest==7.4.0
pytest-cov==4.1.0
